feat(components/molecule/select): remmove withOpenToggle HoC
